requirements.txt
streamlit==1.38.0
pandas==2.2.2
plotly==5.24.1
sqlalchemy==2.0.32
pyodbc==5.1.0
python-dotenv==1.0.1
